# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Book API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/circle-join-requests": {
            "post": {
                "description": "Joins immediately if the link doesn't require approval, otherwise opens a join request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "circle-join-requests"
                ],
                "summary": "Follow a circle invite link",
                "parameters": [
                    {
                        "description": "Invite link token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.FollowInviteLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_FollowInviteLinkResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            }
        },
        "/circle-join-requests/{id}/cancel": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "circle-join-requests"
                ],
                "summary": "Cancel my own circle join request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Join request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_CircleJoinRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            }
        },
...
        "/circles/{id}/join-requests": {
            "get": {
                "description": "The approval queue; requires the caller to have invite privileges",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "circle-join-requests"
                ],
                "summary": "List a circle's pending join requests",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Circle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Requests per page (default 20, max 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_ListCircleJoinRequestsResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            }
        },
        "/circles/{id}/join-requests/{requestId}/approve": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "circle-join-requests"
                ],
                "summary": "Approve a circle join request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Circle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Join request ID",
                        "name": "requestId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_CircleMemberResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            }
        },
        "/circles/{id}/join-requests/{requestId}/reject": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "circle-join-requests"
                ],
                "summary": "Reject a circle join request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Circle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Join request ID",
                        "name": "requestId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_CircleJoinRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.CircleJoinRequestResponse": {
            "type": "object",
            "properties": {
                "circle_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-07-20T10:00:00Z"
                },
                "decided_at": {
                    "type": "string"
                },
                "decided_by_user_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "invite_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                },
                "user_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                }
            }
        },
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.CircleMemberResponse": {
            "type": "object",
            "properties": {
                "can_capture": {
                    "type": "boolean"
                },
                "can_invite": {
                    "type": "boolean"
                },
                "circle_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "joined_at": {
                    "type": "string",
                    "example": "2026-07-20T10:00:00Z"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "user_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.FollowInviteLinkRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.FollowInviteLinkResponse": {
            "type": "object",
            "properties": {
                "join_request": {
                    "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.CircleJoinRequestResponse"
                },
                "member": {
                    "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.CircleMemberResponse"
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.ListCircleJoinRequestsResponse": {
            "type": "object",
            "properties": {
                "join_requests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.CircleJoinRequestResponse"
                    }
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_CircleJoinRequestResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.CircleJoinRequestResponse"
                },
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_CircleMemberResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.CircleMemberResponse"
                },
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_FollowInviteLinkResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.FollowInviteLinkResponse"
                },
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_ListCircleJoinRequestsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.ListCircleJoinRequestsResponse"
                },
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
...
    }
}